requirements.txt
# Core numerical & data handling
numpy>=2.2,<3.0
pandas>=2.3,<3.0
scikit-learn>=1.6,<2.0
scipy>=1.16,<2.0
joblib>=1.5

# Tree-based models
xgboost>=3.0,<4.0

# Deep learning (CPU/GPU handled automatically by PyTorch)
torch>=2.9,<3.0
torchvision>=0.24,<1.0

# Geospatial analysis
geopandas>=1.1,<2.0
shapely>=2.1,<3.0
pyproj>=3.7,<4.0

# Image processing
opencv-python>=4.12,<5.0
Pillow>=12.0,<13.0

# Visualization
matplotlib>=3.10,<4.0
seaborn>=0.13,<1.0

# Utilities
requests>=2.32
tqdm>=4.67
python-dotenv>=1.2

# Notebooks
notebook>=7.5